+++ tbot/machine/channel/channel.py
@@ -299,6 +299,7 @@ class Channel(abc.ABC):
         regex: bool = False,
         stream: typing.Optional[typing.TextIO] = None,
         timeout: typing.Optional[float] = None,
+        must_end: bool = True,
     ) -> str:
         """
         Read until receiving ``prompt``.
@@ -311,13 +312,19 @@ class Channel(abc.ABC):
         :param io.TextIOBase stream: Optional stream where ``read_until_prompt``
             should write everything received up until the prompt is detected.
         :param float timeout: Optional timeout.
+        :param bool must_end: Whether the prompt has to appear at the end of
+            the stream.
         :raises TimeoutError: If a timeout is set and this timeout is reached before
             the prompt is detected.
         :rtype: str
         :returns: Everything read up until the prompt.
         """
         start_time = time.monotonic()
-        expr = f"{prompt}$" if regex else "^$"
+        expr = None
+        if regex:
+            expr = f"{prompt}$" if must_end else prompt
+        elif not must_end:
+            expr = re.escape(prompt)
         buf = ""
 
         timeout_remaining = timeout
@@ -343,13 +350,17 @@ class Channel(abc.ABC):
 
             buf += decoded
 
-            if (not regex and buf[-len(prompt) :] == prompt) or (
-                regex and re.search(expr, buf) is not None
+            if (expr is None and buf[-len(prompt) :] == prompt) or (
+                expr is not None and re.search(expr, buf) is not None
             ):
                 if stream:
-                    stream.write(decoded[: -len(prompt)])
+                    # Don't clip prompt if it doesn't need to be at the end
+                    if must_end:
+                        stream.write(decoded[: -len(prompt)])
+                    else:
+                        stream.write(decoded)
                 break
-            elif stream:
+            elif stream is not None:
                 stream.write(decoded)
 
             if timeout is not None: